Allow font size to be a float (#71)

+++ vl-convert-rs/src/text.rs
@@ -104,7 +104,7 @@ struct TextInfo {
     variant: Option<String>,
     weight: Option<String>,
     family: Option<String>,
-    size: i32,
+    size: f64,
     text: Value,
 }
